plugins/archivist/pyproject.toml
[project]
name = "pkm-plugin"
version = "1.1.0"
description = "Personal Knowledge Management expert for Obsidian vaults"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
